(client): initial commit puddle->transmission interface
This commit adds an initial implementation of the classes needed to
safely interact from the transmission daemon from the puddle frontend.
For now we use webpack-dev-server as a proxy to bypass any CORS
issues. Later on we should introduce a custom backend server which
distributes takes care of the current role of webpack-dev-server.

The interface itself is just a class which forwards requests from
puddle to the transmission RPC interface and then casts the result
to a typescript interface which gives compile time guarantees about
associated fields.