(client): render torrents on the dashboard
Extremely early stage implementation of the dashboard. At the moment
we fetch all the torrents at once at a fixed interval, and then render
them to the DOM. React takes care of optimising the rendering, thereby
maximising speed. But there's much room for improvement here.

The dashboard is implementated using {@code ./src/puddle/torrent-columns}.
That module defines an interface for abstracting the details of a
torrent from the fields needed to calculate it and the order in which
their shown. The docstring is clear enough, so I won't repeat it here.
But suffice it to say that if and when puddle decides to add a new
column, it should be as simply as just creating a new object in the
Columns singleton.

To summarise what still needs to be done:
- Optimise the fetching and rendering of torrents
  - I'm thinking we really don't need to fetch updates for all of them
    at once. Fetching, say, all the ones on the screen and then upto
    some limit above and below should suffice for most use cases. We
    still need to maintain a collection of all the torrents (for
    using) trackers and other data in filtering/naming. But updating
    only a subset of them (or perhaps chunking updates over multiple
    request intervals) will surely improve performance.
- Restructure the dashboard so the table header is fixed to the top.
  - It's annoying to scroll down and then have to scroll up again
    because you forgot which column you're looking at.
- Setup the sort by columns feature. This'll probably involve adding a
  compare option to the Columns interface to allow sorting. But at the
  moment I'm okay with just flipping ascending to descending and
  vice/versa.
- make torrents selectable for operations using the header buttons
- add a right click context menu for more expansive operations.

After that I can start focusing on filtering, settings and deployment
servers.

Stray Changes
- I found out why I couldn't reference @puddle as an import. Turns out
  tsc has its own module aliasing mechanism and webpack ts-checker
  plugin won't use webpacks module paths at all :-(. Fixed it now by
  updating tsconfig.json, the redundancy is annoying but its to be
  expected I suppose.
- Annoyingly it turns out that webpack won't actually process any
  imports if you don't use them on the page. Using them as a type
  doesn't count as using them so some of the code in the responses
  interfaces for our transmission interface had type bugs and webpack
  just didn't decide to let me know. moral of the story, check types
  seperately from development :-).