(client): switch to redux for state management
It was becoming a real pain to manage application state from a single
component and then pass it down through to other components. Firstly
it makes components exttremely ugly, with shoehorned properties and
inline lambdas designed to invoke some setter somewhere higher up in
the component hirearchy. Secondly I had to keep defining and modifying
interfaces to make inter component communications work. This patch
switches to redux, which is a far more sane approach to state
management. Sadly it's broken torrent filtering, but selection and
display is still working :smile:.