(client): restructure store to optimize rendering
So far I've been storing torrents as a collection, and then selected
torrents (on the dashboard) as an array of torrent ids.

This was easy to implement and allowed me to avoid defining a seperate
type for responses from the transmission API and our own store
interface. However this had the annoying consequence of me having to
request the selected torrents in the render method for the dashboards
table body which resulted in all the rows in the table being
re-rendered every time you selected one.