(client): add support for filtering by tracker/status
This patch reapplies support for filtering the displayed torrent list
by the status of the torrent or trackers.

At the moment the filtering is **and**ing across each filter, while
**or**ing within the same filter. Meaning if you filter to only show
Active torrents, and then by trackers foo and bar. You'll get a list
of all active torrents that share either foo or bar as a tracker.