(client): make progress icon match state
Up until know I've just been using a static ! icon for demo purposes,
but now the icon should match the status of the torrent in question.

At the moment it priroritises errors first, then whether the torrent
is paused, whether it's actively downloading or seeding and otherwise
the torrent can only possibly be queued.

I'm not sure whether there's any merit to distinguish between seeding
and downloading torrents. I'll come back to that.