(client): add sidebar button to toggle alt-speed
it's a little turtle. if you click it transmission will switch to
using its alternative speed.

At the moment, when it's active, it looks the same as if you're
hovering over it. I'm considering just showing a little disc in the
top right corner instead, if nothing else it would make things less
ambiguous.