(client): add login, loading and client views
Now the user will be presented a login screen on startup, from which
they should enter the transmission URL and then be redirected to the
dashboard. For now I'm bypassing the screen because it'll too much of
a pain to resubmit every time I refresh the page. Once we have user
sessions that won't really still be an issue.

Later on we should switch from requesting transmission details to
logging in actual user accounts.