(client): fix column resize is broken on scroll
If you scroll the table body slightly to the right, the column resize
functionality was broken :-(, because the minimum position of the
column was being calculated as if all the columns were visible instead
of just the ones from the scroll point.

This patch also fixes the issue where moving the mouse in the
torrents-list while resizing doesn't move the resize element. I'm not
completely sure why, but moving the resize overlay to the end of the
torrents list fixed it :-).