(client): improve keyboard based accessibility
Attach a root key handler to the dashboard view which monitors and
overrides some key combinations to interact with the dashboard. So
far we've got:

| Key     | Command                        |
|---------|--------------------------------|
| C-g     | escape, refocus the root div   |
| C-s/C-f | switch focus to the search bar |
| C-a     | select all torrents            |
| C-d     | deselect all torrents          |

The following are available exclusively in the search bar:

| Key    | Command                              |
|--------|--------------------------------------|
| C-k    | cancel this search                   |
| escape | switch focus to last focused element |

I've also attached tabIndex properties to most selectable
elements (action buttons, tooltip buttons, the search bar
and filter lists) meaning you can tab to them and use enter
to perform the same behaviour as a click.

This patch involved making some slight modifications to the
clientview (because it was getting too long). I've moved some
core logic into some other files in the same directory.