(client): initial commit notifications framework
now you get tiny notifications in the bottom right hand corner to
alert you when something happens. At the moment there're just
notifications for when a torrent is added, and when a torrent is
removed.

Later on I'm hoping to be able to use them to maintain a backlog of
events for the transmission daemon. The idea being that I can remove a
torrent... later decide I didn't want to do that and then look back
through my notifications to find where it says the torrent was removed
and it's magnet link (or hash) and then re-add it. There should be a
limit variable for it as well, but seeing as (atm) notifications are
erased on reload, that's not a big concern.