(client): switch to momentjs for time formatting
It's seriously insane how there's no built in datetime format that
isn't a pain to use. momentjs exists so it is possible. That said,
date time is a serious pain so its understandable no browser or
standard is willing to incorporate some interface which it'll probably
deprecate 2 years down the line (ahhmmm, ahhhmmmm java).