(bug): don't store updater canceler in component state
Turns out I was overthinking this, there's no need (and actually quite
a few reasons not) to store the updater canceler in the client state.

Firstly it causes a double render whenever the components is rendered.
Secondly I can just return the canceler from the effect call and react
will guarantee the previous updaters are cancelled before the new ones
are assigned.