(bug): changing updater intervals sometimes doesn't clean up
Specifically when you clear a timoue at the exact moment a new timeout
is assigned. I'm quite sure its happening between when we check an
interval is assigned and then clear it. The two operations aren't
atomic, which is what was causing the issue.

I've switched to an external package to maintain the state of
intervals and now we're using a class to abstract the starting and
stopping of asynchronous update logic.