(cleanup): move styles out of components directory
At first I thought this would be easier... but it's really ugly to
find styles mixed in with scripts and it's hard to find styles using
my editor because the only common part of filenames is the extension,
not the prefix.

Now I've just moved everything to the styles directory. That said I
have kept some small or loosely isolated styles in the components
directory, just because the effort of specifying a longer import path
didn't seem to warant how closely related the styles are.