(client): add overlays/modals support
Now there's a framework for spawning new overlays/container elements
on top of the dashboard. These overlays can be exited by clicking off
the side of the modal or each modal can do so by dispatching to the UI
store.

For now I've added a torrent details and a settings modal, but all
they do is show a title. Now it's time to fill them out :smile:.