(client): initial commit torrent details overlay
This technically isn't ready yet, but there's a lot of changes and it
seems like a good time to commit what I've gotten so far.

So now you can click on the torrent-details button in the context menu
to pop open an overlay. From the overlay you can see peer data, set the
torrents bandwidth priority, start and stop it, get general information
(such as the total torrent size or creator), peer and tracker data and
at the moment a static file list. I was hoping to use the file list to
set the priority for individual files but reacts Context API is giving
me performance issues :cry:. I'll have to switch to storing view context
in the redux store to optimise performance, it'll be a pain to get
around typescripts nullability warnings but the performance gains are
worth it.