(client): move overlays state to redux store
The start of creating real performance improvements for the overlay
subsystem. At the moment I've just moved the stuff I was storing in
a single root object into a shared subsection of the redux store. I
would've liked to keep it as a subpart of the UI store, but it doesn't
seem like there's a way to combineReducers without having to prefix
them with their own unique keys. So I would've had to have ui.props
for the stuff related to all ui components, and ui.torrent-details for
stuff just for the torrent-details store... but the props prefix is
meaningless so I don't wanna have to specify it :tongue:. In the end
we've got ui.overlay and overlays.torrent-details which seems like a
decent compromise.

Either way, next step is to move torrent-full to the stores
directory... which is getting kinda full to be honest. It might be
worth refactoring it into a models directory and a stores directory
respectively.