(client): support changing file priority from details view
Now there's a slider (just like there is for bandwidth priority) on
the file list which you can click to toggle the bandwidth priority of
a torrent. By default transmission sets a priority of Norm, so it'll
go Norm -> High -> Don't Download -> Low -> Norm.

The combobox approach for bulk toggling the priority is still not
implemented because I'm considering switching to a dedicated component
for it.