(client): add feature, set torrent labels
The context menu item for settings tags now pops open an overlay from
which you can select multiple tags to assign to the selected torrents.

There's completion offered for the list of existing tags as well.

I've also cleaned up my inconsistent use of tags and labels. In the
implementation we show them (to the user) as tags because that's a
more intuitive explanation of their purpose. In the code we write it
as labels because that's what transmission calls it.