(repo): fix jest crashes when importing styles
Jest didn't have any of the module mappings from webpack/tsc so I had
to add them in, and then added a proxy module for stylesheets so it
doesn't crash whenever I import a component.

I've had to respecify the module maps three times now. Node should
really have a framework for specifying them in one file, I'd do so
but some of the configs are in JSON, not JS so repetition would be
inevitable :/.

I've also added some tests for cycle-priority here.