(server): add user authentication through sessions
Now users are presented with a login form on initial connection, from
there they enter a transmission URL. Once the URL is verified its
stored in the backend as a session and any requests to the servers
transmission end point proxies through the server to that URL. Users
can logout by having this session data erased.

I've also created an API endpoint to discover what the default URL for
transmission would be, based on the users transmission configuration
and this is what's presented to the user on the logon form.

Lastly I've added hot-reload support to the webpack configuration.
Honestly I just wanted the page to auto-reload whenever I make a
change but it seems like that alone isn't an option independent from
hot reloading :/.