(client): add persistent notifications and notification details
Now, a subset of dispatched notifications are transmitted to and
stored in the backend server. The client has an overlay to inspect
displayed notifications and it allows you to fetch excess notification
from the servers archive.

Each notification in the overlay can be clicked to show more details
about the event. So far you can see some metadata about added or
removed torrents. Other notification types are pending.