(tests): increase test coverage
Now a good portion of the complex parts of the redux store has test
coverage, I've also added snapshot tests for ~~a~~ some components and
tried using mock-fetch to fake some API calls in redux thunks. I'd
like to add more test coverage, but its a slow and gruelling process
and I'm running low on time. Hopefully we'll have full test coverage
by V2.

I've also moved the API module from the client package to the server
package. In the future I'm planning to add typescript interfaces for
request and response types so that I can have compile time safety for
api handlers. For now the Server is so minimal it's not a big concern,
but by V2 it'll be almost essential.