(repo): setup documentation build action
This patch improves automatic documentation rendering. Now any push to
master will trigger a rebuilding of the documentation, followed by a
push to the github-pages branch of the remote repository so that
rendered documentation is immeadiately accessible.

I've also switched doc themes to minami (mostly because the default
theme) isn't very easy on the eyes or to traverse.